+++ src/main/java/org/scion/jpan/internal/paths/PathBuilder.java
@@ -16,6 +16,8 @@ package org.scion.jpan.internal.paths;
 
 import com.google.protobuf.ByteString;
 import com.google.protobuf.InvalidProtocolBufferException;
+
+import java.net.InetSocketAddress;
 import java.nio.ByteBuffer;
 import java.util.*;
 import java.util.stream.Collectors;
@@ -24,6 +26,7 @@ import org.scion.jpan.ScionRuntimeException;
 import org.scion.jpan.ScionUtil;
 import org.scion.jpan.internal.bootstrap.LocalAS;
 import org.scion.jpan.internal.util.ByteUtil;
+import org.scion.jpan.internal.util.IPHelper;
 import org.scion.jpan.internal.util.MultiMap;
 import org.scion.jpan.proto.control_plane.Seg;
 import org.scion.jpan.proto.crypto.Signed;
@@ -394,9 +397,9 @@ public class PathBuilder {
     path.setRaw(raw);
 
     // First hop
-    String firstHop =
-        localAS.getFirstHopAddressString((int) path.getInterfaces().get(0).getId());
-    path.setLocalInterface(PathMetadata.Interface.create(firstHop));
+    int firstHopId = (int) path.getInterfaces().get(0).getId();
+    InetSocketAddress firstHop = localAS.getFirstHopAddress(firstHopId);
+    path.setLocalInterface(PathMetadata.Interface.create(IPHelper.toString(firstHop)));
 
     // Metadata
     SegmentMetadataAccumulator.writeStaticInfoMetadata(path, segments, ranges);